chore: add findDOMNode polyfill
